shirts hooked up to the database
